Move token tests to separate file
